requirements.txt
django==4.2.2
djangorestframework==3.14.0 
pandas==2.0.2
certifi==2020.6.20
click==7.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
numpy==1.22.0
scikit-learn==1.0.2
scipy==1.10.1
sklearn==0.0.post7
threadpoolctl==3.1.0
matplotlib==3.7.1
seaborn==0.12.2
statsmodels==0.14.0
uvicorn==0.22.0
apscheduler